fixed 32bit int file size limit

+++ src/api/MediaServer1ItemRes.h
@@ -50,7 +50,7 @@
     NSString *duration;
     int nrAudioChannels;
     NSString *protocolInfo;
-    int size;
+    long long int size;
     int durationInSeconds;
 }
 
@@ -60,7 +60,7 @@
 @property(retain, nonatomic) NSString *duration;
 @property(readwrite) int nrAudioChannels;
 @property(retain, nonatomic) NSString *protocolInfo;
-@property(readwrite) int size;
+@property(readwrite) long long int size;
 @property(readwrite) int durationInSeconds;
